pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "awesome-confluence-mcp"
version = "1.0.0"
description = "Token-efficient Confluence MCP server that converts pages to Markdown"
authors = [
    { name = "mazhar480", email = "your.email@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["mcp", "confluence", "markdown", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "markdownify>=0.11.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/mazhar480/awesome-confluence-mcp"
Repository = "https://github.com/mazhar480/awesome-confluence-mcp"
Issues = "https://github.com/mazhar480/awesome-confluence-mcp/issues"

[project.scripts]
awesome-confluence-mcp = "server:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
